feat(config): add new setting to set site/system name in the pretalx.cfg file
This commit adds a new setting to the pretalx.cfg file to set the site/system name. This setting is used in the site header and in the page title.
This feat also sets the deault site name to "Eventyay" in the CONFIG constant under config.py and sets the sample in the pretalx.example.cfg file.
This change updates the tempalte pretalx/orga/templates/orga/base.html to use the new setting to set the site name in the site header and page title.